+++ activesupport/CHANGELOG.md
@@ -274,4 +274,8 @@
 
     *Jonathan del Strother*
 
+*   `ActiveSupport::TimeWithZone# utc?` to remove `TZ='UCT'` and add `TZ='GMT'.
+
+    *Haruki Kobayashi*
+
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/activesupport/CHANGELOG.md) for previous changes.